% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{forint}
\alias{forint}
\title{Formats a number to HUF}
\usage{
forint(x)
}
\arguments{
\item{x}{numeric vector, pereferably a monetary value}
}
\value{
string formated for ease of use
}
\description{
Formats a number to HUF
}
\examples{
forint(42)
forint(420189567.244)
forint(1:10)
}
