% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bitcoin_price.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Gets the latest bitcoin price in USD}
\usage{
get_bitcoin_price(retried = 0)
}
\arguments{
\item{retried}{it defines the system's sleeping time. No input needed.}
}
\value{
numeric the current Bitcoin price
}
\description{
Gets the latest bitcoin price in USD
}
\examples{
get_bitcoin_price()
}
