% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rates.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Gets historical exchange rates}
\usage{
get_exchange_rates(
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  retried = 0,
  numerator = "USD",
  denominator = "HUF"
)
}
\arguments{
\item{start_date}{starting date to receive exchange rates information}

\item{end_date}{the last date for our exchange rates}

\item{retried}{it defines the system's sleeping time. No input needed.}

\item{numerator}{numerator of the requested exchange rate}

\item{denominator}{denominator of the requested exchange rate}
}
\value{
\code{data.table} object
}
\description{
Gets historical exchange rates
}
\examples{
get_exchange_rates(start_date = "2021-05-16", end_date = "2021-05-18", numerator = "EUR", denominator = "USD")
}
