% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usdhuf.R
\name{get_usdhuf}
\alias{get_usdhuf}
\title{Gets the latest USD/HUF rate}
\usage{
get_usdhuf(retried = 0)
}
\arguments{
\item{retried}{it defines the system's sleeping time. No input needed.}
}
\value{
numeric the current USD/HUF rate
}
\description{
Gets the latest USD/HUF rate
}
\examples{
get_usdhuf()
}
