% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usdhuf.R
\name{get_usdhufs}
\alias{get_usdhufs}
\title{Gets the historical USD/HUF rate}
\usage{
get_usdhufs(start_date = Sys.Date() - 30, end_date = Sys.Date(), retried = 0)
}
\arguments{
\item{start_date}{starting date to receive exchange rates information}

\item{end_date}{the last date for our exchange rates}

\item{retried}{it defines the system's sleeping time. No input needed.}
}
\value{
\code{data.table} object
}
\description{
Gets the historical USD/HUF rate
}
\examples{
get_usdhufs(start_date = "2021-05-16", end_date = "2021-05-18")
}
